:recycle: 모듈 역할 기반 분리
