+++ app/blueprints/standard/standard_blueprint.py
@@ -17,7 +17,8 @@ from app.schemas.success_response import SuccessResponse
 from app.services.common.ingestion_pipeline import preprocess_data, \
   chunk_standard_texts, normalize_spacing
 from app.services.standard.vector_delete import delete_by_standard_id
-from app.services.standard.vector_store import vectorize_and_save
+from app.services.standard.vector_store.vector_processor import \
+  vectorize_and_save
 
 standards = Blueprint('standards', __name__, url_prefix="/flask/standards")
