--- app/services/standard/vector_store.py
@@ -1,170 +1,0 @@
-import asyncio
-import logging
-import time
-import uuid
-from collections.abc import Callable
-from datetime import datetime
-from typing import List, Coroutine, Any
-
-import numpy as np
-from httpx import ConnectTimeout
-from qdrant_client import AsyncQdrantClient
-from qdrant_client.http.exceptions import ResponseHandlingException
-from qdrant_client.models import Distance, VectorParams, PointStruct
-
-from app.blueprints.standard.standard_exception import StandardException
-from app.clients.openai_clients import get_prompt_async_client, \
-  get_embedding_async_client
-from app.clients.qdrant_client import get_qdrant_client
-from app.common.constants import MAX_RETRIES, LLM_TIMEOUT
-from app.common.exception.custom_exception import CommonException
-from app.common.exception.error_code import ErrorCode
-from app.containers.service_container import embedding_service, prompt_service
-from app.models.vector import VectorPayload, WordPayload
-from app.schemas.chunk_schema import ClauseChunk
-from app.schemas.document_request import DocumentRequest
-
-STANDARD_LLM_REQUIRED_KEYS = {"incorrect_text", "corrected_text"}
-STANDARD_WORD_REQUIRED_KEYS = {"keyword", "meaning_difference"}
-
-
-async def vectorize_and_save(chunks: List[ClauseChunk],
-    pdf_request: DocumentRequest) -> None:
-  qd_client = get_qdrant_client()
-  await ensure_qdrant_collection(qd_client, pdf_request.categoryName)
-
-  semaphore = asyncio.Semaphore(5)
-  async with get_prompt_async_client() as prompt_client:
-    task_fn = make_word_payload if pdf_request.categoryName == "법률용어" else make_clause_payload
-    results = await asyncio.gather(*[
-      task_fn(prompt_client, article, pdf_request, semaphore)
-      for article in chunks
-    ])
-
-  embedding_inputs = [point.embedding_input() for point in results if point]
-
-  async with get_embedding_async_client() as embedding_client:
-    start_time = time.perf_counter()
-    embeddings = await embedding_service.batch_embed_texts(embedding_client,
-                                                           embedding_inputs)
-    logging.info(f"임베딩 묶음 소요 시간: {time.perf_counter() - start_time:.4f}초")
-
-  final_points = []
-  for payload, vector in zip(results, embeddings):
-    if vector and isinstance(vector, list) and not any(
-        np.isnan(x) for x in vector):
-      final_points.append(build_point(payload, vector))
-
-  await upload_points_to_qdrant(qd_client, pdf_request.categoryName,
-                                final_points)
-
-
-async def make_clause_payload(prompt_client, article, pdf_request,
-    semaphore) -> VectorPayload | None:
-  async with semaphore:
-    result = await retry_llm_call(
-        prompt_service.make_correction_data,
-        prompt_client, article,
-        required_keys=STANDARD_LLM_REQUIRED_KEYS
-    )
-    if not result:
-      return None
-
-  return VectorPayload(
-      standard_id=pdf_request.id,
-      incorrect_text=result.get("incorrect_text") or "",
-      proof_text=article.clause_content,
-      corrected_text=result.get("corrected_text") or "",
-      created_at=datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-  )
-
-
-async def make_word_payload(prompt_client, article: ClauseChunk, pdf_request,
-    semaphore) -> WordPayload | None:
-  async with semaphore:
-    result = await retry_llm_call(
-        prompt_service.extract_keywords,
-        prompt_client, article.clause_content,
-        required_keys=STANDARD_WORD_REQUIRED_KEYS
-    )
-    if not result:
-      return None
-
-  return WordPayload(
-      standard_id=pdf_request.id,
-      definition=article.clause_content,
-      term=article.clause_number,
-      meaning_difference=result.get("meaning_difference") or "",
-      keywords=result.get("keyword") or [],
-      created_at=datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-  )
-
-
-def build_point(payload, embedding: List[float]) -> PointStruct:
-  return PointStruct(
-      id=str(uuid.uuid4()),
-      vector=embedding,
-      payload=payload.to_dict()
-  )
-
-
-async def retry_llm_call(
-    func: Callable[..., Coroutine[Any, Any, dict]],
-    *args,
-    required_keys: set | None = None) -> dict | None:
-  for attempt in range(1, MAX_RETRIES + 1):
-    try:
-      result = await asyncio.wait_for(func(*args), timeout=LLM_TIMEOUT)
-      if isinstance(result, dict) or required_keys.issubset(result.keys()):
-        return result
-      logging.warning(
-          "[retry_make_correction]: llm 응답 필수 키 누락 / dict 구조 아님")
-
-    except asyncio.TimeoutError:
-      logging.warning(
-        f"[retry_make_correction]: Timeout {attempt}/{MAX_RETRIES}")
-      if attempt == MAX_RETRIES:
-        raise CommonException(ErrorCode.LLM_RESPONSE_TIMEOUT)
-
-    except Exception as e:
-      logging.warning(
-          f"[retry_make_correction]: 재요청 발생 {attempt}/{MAX_RETRIES} {e}")
-      if attempt == MAX_RETRIES:
-        raise StandardException(ErrorCode.STANDARD_REVIEW_FAIL)
-
-    await asyncio.sleep(0.5 * attempt)
-
-  raise StandardException(ErrorCode.PROMPT_MAX_TRIAL_FAILED)
-
-
-async def ensure_qdrant_collection(qd_client: AsyncQdrantClient,
-    collection_name: str) -> None:
-  try:
-    exists = await qd_client.collection_exists(collection_name=collection_name)
-    if not exists:
-      await create_qdrant_collection(qd_client, collection_name)
-
-  except (ConnectTimeout, ResponseHandlingException):
-    raise CommonException(ErrorCode.QDRANT_NOT_STARTED)
-
-
-async def create_qdrant_collection(qd_client: AsyncQdrantClient,
-    collection_name: str):
-  try:
-    return await qd_client.create_collection(
-        collection_name=collection_name,
-        vectors_config=VectorParams(size=1536, distance=Distance.COSINE)
-    )
-  except (ConnectTimeout, ResponseHandlingException):
-    raise CommonException(ErrorCode.QDRANT_CONNECTION_TIMEOUT)
-
-
-async def upload_points_to_qdrant(qd_client: AsyncQdrantClient, collection_name,
-    points):
-  if not points:
-    raise StandardException(ErrorCode.NO_POINTS_GENERATED)
-
-  try:
-    await qd_client.upsert(collection_name=collection_name, points=points)
-  except (ConnectTimeout, ResponseHandlingException):
-    raise CommonException(ErrorCode.QDRANT_CONNECTION_TIMEOUT)
